+++ terraform/key-vault.tf
@@ -14,10 +14,10 @@ module "key_vault" {
 
 import {
   to = module.key_vault.azurerm_key_vault.vault
-  id = "/subscriptions/28aefbe7-e2af-4b4a-9ce1-92d6672c31bd/resourceGroups/rg-ytsumm-prd/providers/Microsoft.KeyVault/vaults/kv-ytsumm-prd"
+  id = "/subscriptions/28aefbe7-e2af-4b4a-9ce1-92d6672c31bd/resourceGroups/rg-ytsumm-prd-ci/providers/Microsoft.KeyVault/vaults/kv-ytsumm-prd-ci"
 }
 
 import {
   to = module.key_vault.azurerm_role_assignment.secrets_officer[0]
-  id = "/subscriptions/28aefbe7-e2af-4b4a-9ce1-92d6672c31bd/resourceGroups/rg-ytsumm-prd/providers/Microsoft.KeyVault/vaults/kv-ytsumm-prd/providers/Microsoft.Authorization/roleAssignments/fde6c717-0024-8ad6-efd0-ce353b6928f3"
+  id = "/subscriptions/28aefbe7-e2af-4b4a-9ce1-92d6672c31bd/resourceGroups/rg-ytsumm-prd-ci/providers/Microsoft.KeyVault/vaults/kv-ytsumm-prd-ci/providers/Microsoft.Authorization/roleAssignments/9bdcfc58-afc3-4278-b94b-17f2d5f90aed"
 }