+++ terraform/main.tf
@@ -10,5 +10,5 @@ resource "azurerm_resource_group" "main" {
 
 import {
   to = azurerm_resource_group.main
-  id = "/subscriptions/28aefbe7-e2af-4b4a-9ce1-92d6672c31bd/resourceGroups/rg-ytsumm-prd"
+  id = "/subscriptions/28aefbe7-e2af-4b4a-9ce1-92d6672c31bd/resourceGroups/rg-ytsumm-prd-ci"
 }